+++ src/main/java/domain/ladder/Ladder.java
@@ -7,16 +7,8 @@ import java.util.List;
 public class Ladder {
     private final List<Floor> ladder;
 
-    private Ladder(List<Floor> ladder) {
+    public Ladder(List<Floor> ladder) {
         this.ladder = ladder;
-    }
-
-    public static Ladder createByStrategy(BridgeGenerator bridgeGenerator, Height height, Width width) {
-        final List<Floor> ladder = new ArrayList<>();
-        for (int i = 0; i < height.getValue(); i++) {
-            ladder.add(FloorFactory.createByStrategy(bridgeGenerator, width));
-        }
-        return new Ladder(ladder);
     }
 
     public int calculateResultPosition(final int startPosition) {
